Commit before refactor list result
Rather than build a proplist of name and value for a file wildcard
match, we return a list. This idea of having an implicit 'name' for
each file is bad. We have the item itself, marked with __file__, and
then use filters to find that item, either by file name or some other
attribute.