Missing item context name
By default, the template_map item context is named 'item'. For
readability however, you can specify a different name by using a
two-tuple of {Name, List} in place of the List.

Lambda Pad needs better validation (a warning, e.g.) in this case
as the target names will all colide to "site/.html", which is
something we can detect and complain about.

The most glaring problem though is that docs for the data loaders
and generators are weak to begin with an incomplete. An idea: use
Erlang specs to generate nice looking docs (not edocs, or edocs with
decent styling). This is definitely a vision for Lambda Pad -
supporting language aware doc generators. The first pass at this
should probably be Lambda Pad itself.