Two new filters: filename and basename
These can be applied to content that is decorated with a __file__ attribute
to get the filename and the basename.

The most common use case for this is to use basename for use in naming
target files for template maps. E.g. "site/{{item|basename}}.html" would
reuse the basename of the source file along with the html extension.