Added a render filter
This can be used to render a string or file reference as a template.
The render does not provide any context and so is currently limited
to conditional logic and includes (includes not tested).

The immediate application of this is to support template logic within
markdown files. Note again however there is not context provided to
the template and is therefore somewhat limited.