Added strip_p filter
This removes leading and trailing 'p' elements, which is
useful for filtering HTML generated from markdown when the
markdown is a fragment.

This should arguably be a different filter (e.g.
markdown_fragment_to_html) but works for now. Let's see how it
goes and modify later if it makes sense.