Handling malformed template compile error
If an include file doesn't exit our version of erlydtl gives us
an error that doesn't include line or col info - this cases handles
it and does the best by putting the user on the last line of the
file (hacked using a very high number).

When we move to a later version of erlydtl we can clean this up.